Static analysis code remediation: ome-xml and specification Java sources

+++ ome-xml/src/main/java/ome/xml/model/primitives/PositiveFloat.java
@@ -44,7 +44,7 @@ public class PositiveFloat extends NonNegativeFloat {
 
   public PositiveFloat(Double value) {
     super(value);
-    if (value == null || value.doubleValue() <= 0) {
+    if (value == null || value <= 0) {
       throw new IllegalArgumentException(
           value + " must be non-null and strictly positive.");
     }
